clean up integratedAM and goal getter
